move const.c and function.c to variable.c
